requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.8.1
asttokens==2.2.1
backcall==0.2.0
bcrypt==5.0.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==3.1.1
colorama==0.4.6
comm==0.1.4
comtypes==1.1.14
contourpy==1.2.0
cryptography==46.0.3
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
ecdsa==0.19.1
executing==1.2.0
fastapi==0.120.0
Flask==2.2.3
fonttools==4.49.0
git-filter-repo==2.47.0
greenlet==3.2.4
groq==0.33.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.4
ipykernel==6.25.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.4.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.5
lightgbm==4.6.0
MarkupSafe==2.1.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
MouseInfo==0.1.3
mysqlclient==2.2.7
nest-asyncio==1.5.7
networkx==3.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==23.1
pandas==2.3.3
parso==0.8.3
passlib==1.7.4
pgzero==1.2.1
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2-binary==2.9.11
pure-eval==0.2.2
pyasn1==0.6.1
PyAutoGUI==0.9.53
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.15.1
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pyttsx3==2.90
pytweening==1.0.4
pytz==2024.1
pywhatkit==5.4
pywin32==305
PyYAML==6.0.3
pyzmq==25.1.0
reportlab==4.4.0
requests==2.28.2
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.4
SQLAlchemy==2.0.44
sqlparse==0.5.3
stack-data==0.6.2
starlette==0.48.0
threadpoolctl==3.6.0
tornado==6.3.2
traitlets==5.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
urllib3==1.26.15
uvicorn==0.38.0
watchfiles==1.1.1
wcwidth==0.2.6
websockets==15.0.1
Werkzeug==2.2.3
wikipedia==1.4.0
